requirements.txt
jsonrpclib-pelix
lxml
nltk
numpy
pandas
pika
pymongo
python-dateutil
redis
requests
scipy
sklearn
tensorflow
watchdog